Improvement, update integration tests
