+++ package.json
@@ -28,7 +28,7 @@
   "bin": {},
   "scripts": {
     "test": "nyc mocha -r ts-node/register --project tsconfig.tests.json test/unit-tests/index.test.ts && nyc report --reporter=text-summary",
-    "integration-test": "nyc mocha -r ts-node/register --project tsconfig.tests.json test/integration-tests/integration.test.ts && nyc report --reporter=text-summary",
+    "integration-test": "nyc mocha -r ts-node/register --project tsconfig.tests.json test/integration-tests/*.test.ts && nyc report --reporter=text-summary",
     "lint": "tslint --project . && tslint --project tsconfig.tests.json",
     "build": "tsc --project ."
   },
