--- test/integration-tests/disabled/serverless.yml
@@ -1,24 +1,0 @@
-# Disabled
-service: domain-manager-disabled-${opt:RANDOM_STRING}
-provider:
-  name: aws
-  runtime: nodejs12.x
-  region: us-west-2
-functions:
-  helloWorld:
-    handler: handler.helloWorld
-    events:
-      - http:
-          path: hello-world
-          method: get
-          cors: true
-plugins:
-  - serverless-domain-manager
-custom:
-  customDomain:
-    domainName: disabled-${opt:RANDOM_STRING}.${env:TEST_DOMAIN}
-    enabled: false
-
-package:
-  exclude:
-    - node_modules/**
