+++ test/integration-tests/generic/apigateway-with-export/handler.js
