+++ test/integration-tests/generic/apigateway-with-import/handler.js
