+++ test/integration-tests/generic/apigateway-with-import/serverless.yml
@@ -5,7 +5,7 @@ provider:
   name: aws
   runtime: nodejs12.x
   region: us-west-2
-  stage: dev
+  stage: test
   apiGateway:
     restApiId:
       Fn::ImportValue: api-${self:provider.stage}-restApiId-${opt:RANDOM_STRING}
