+++ test/integration-tests/generic/basepath-mapping/handler.js
