+++ test/integration-tests/generic/basepath-mapping/serverless.yml
