+++ test/integration-tests/generic/create-domain-idempotent/serverless.yml
