+++ test/integration-tests/generic/null-basepath-mapping/handler.js
