+++ test/integration-tests/generic/null-basepath-mapping/serverless.yml
