--- test/integration-tests/regional-tls-1-0/serverless.yml
@@ -1,28 +1,0 @@
-# Test regional domains with TLS 1.0
-service: regional-tls-1-0-${opt:RANDOM_STRING}
-provider:
-  name: aws
-  runtime: nodejs12.x
-  region: us-west-2
-  stage: dev
-  endpointType: regional
-
-functions:
-  helloWorld:
-    handler: handler.helloWorld
-    events:
-      - http:
-          path: hello-world
-          method: get
-          cors: true
-plugins:
-  - serverless-domain-manager
-custom:
-  customDomain:
-    domainName: regional-tls-1-0-${opt:RANDOM_STRING}.${env:TEST_DOMAIN}
-    securityPolicy: tls_1_0
-    endpointType: regional
-
-package:
-  exclude:
-    - node_modules/**
