+++ test/integration-tests/single-cases/auto-domain/handler.js
