+++ test/integration-tests/single-cases/auto-domain/serverless.yml
