+++ test/integration-tests/single-cases/enabled-custom-apigateway/handler.js
