+++ test/integration-tests/single-cases/enabled-custom-apigateway/serverless.yml
