+++ test/integration-tests/single-cases/enabled-regional-basepath/handler.js
