+++ test/integration-tests/single-cases/enabled-regional-basepath/serverless.yml
