+++ test/integration-tests/single-cases/enabled-regional-stage-basepath/handler.js
