+++ test/integration-tests/single-cases/enabled-regional-stage-basepath/serverless.yml
