+++ test/integration-tests/single-cases/enabled-stage-basepath/handler.js
