+++ test/integration-tests/single-cases/enabled-stage-basepath/serverless.yml
