+++ test/integration-tests/single-cases/web-socket/handler.js
