+++ test/integration-tests/single-cases/web-socket/serverless.yml
@@ -3,7 +3,7 @@ provider:
   name: aws
   runtime: nodejs12.x
   region: us-west-2
-  stage: dev
+  stage: test
 functions:
   connect:
     handler: handler.connect
